requirements_part2.txt
Django==5.2.7
psycopg2-binary==2.9.9
djangorestframework==3.15.2
